+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Changed
+
+- `bill`: All invoices will now have a default set of `$tags` that can be used.
+
 ## [v0.219.0] - 2025-07-08
 
 ### Changed